+++ src/app/locations/components/table-locations/table-locations.component.html
@@ -1,3 +1,4 @@
+<button (click)="redirect()">Atras</button>
 <table>
     <tr>
         <th>Name</th>
