+++ src/app/locations/services/locations.service.ts
@@ -2,7 +2,7 @@ import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { Observable } from 'rxjs';
 import { ILocationsComplete } from '../interfaces/ilocations-complete';
-
+import { Router } from '@angular/router';
 @Injectable({
   providedIn: 'root'
 })
